requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.7
django-allauth==0.41.0
django-crispy-forms==1.9.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.21
